Fix logo background image scaling improperly
Body height was set to 100vh, which meant containers (including the
background image's container) were restricted to that height as well.
When page height became greater than viewport height, containers were
unable to scale with it.

To fix the issue, min-height was used to set 100vh instead. This ensures
that while the body always takes up at least 100% of the viewport's
height, its containers can scale up to fit content when needed.